+++ test/starkbank/test_transfer.rb
@@ -20,14 +20,23 @@ describe(StarkBank::Transfer, '#transfer#') do
     File.binwrite('transfer.pdf', pdf)
   end
 
-  def example
+  it 'create and delete' do
+    transfer = example(schedule: true)
+    transfer = StarkBank::Transfer.create([transfer])[0]
+    delete_transfer = StarkBank::Transfer.delete(transfer.id)
+    expect(transfer.id).must_equal(delete_transfer.id)
+    expect(delete_transfer.status).must_equal('canceled')
+  end
+
+  def example(schedule: false)
     StarkBank::Transfer.new(
       amount: rand(1000),
       name: 'João',
       tax_id: '01234567890',
       bank_code: '01',
       branch_code: '0001',
-      account_number: '10000-0'
+      account_number: '10000-0',
+      scheduled: schedule ? Time.now + 24 * 3600 : nil
     )
   end
 end